Drop a redundant widening in the churn count
`count_ones` already answers in u32, so converting it to u32 was a no-op that
clippy fails the build over. Caught by CI rather than locally, because the
workspace lint runs with `--all-targets` and the desktop crate cannot be built
on this machine at all -- its GTK and WebKit dependencies are not installed, so
`cargo clippy --workspace` never reaches it here.

+++ apps/desktop/src-tauri/src/lib.rs
@@ -1031,7 +1031,7 @@ impl DisplayController {
         let changed: u32 = last
             .iter()
             .zip(packed)
-            .map(|(before, after)| u32::from((before ^ after).count_ones()))
+            .map(|(before, after)| (before ^ after).count_ones())
             .sum();
         #[expect(
             clippy::cast_precision_loss,
